[Spring Boot] REST API 게시판 만들기(8) - Service 구현 및 테스트  - BoardServiceTest.java 수정

+++ src/test/java/com/api/board/service/BoardServiceTest.java
@@ -5,6 +5,7 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.List;
 
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
@@ -21,78 +22,95 @@ import com.api.board.domain.Board;
 @RunWith(SpringRunner.class)
 @SpringBootTest
 public class BoardServiceTest {
- 
-    Logger logger = LoggerFactory.getLogger(BoardServiceTest.class);
- 
-    @Autowired
-    BoardService boardService;
- 
-    /** 게시글 목록 조회 시 NULL 아니면 테스트 통과 */
-    @Test
-	public void testGetBoardList() {
-    	 try {
-    		 
-             List<Board> boardList = boardService.getBoardList();
-             assertNotNull(boardList.size());
- 			
-         } catch (Exception e) {
-             e.printStackTrace();
-         }
+
+	Logger logger = LoggerFactory.getLogger(BoardServiceTest.class);
+
+	@Autowired
+	BoardService boardService;
+
+	int boardSeq = 0;
+
+	/** 게시글 목록 조회 시 첫 번째 board_seq 조회 */
+	@Before
+	public void testGetBoardSeq() throws Exception {
+
+		List<Board> boardList = boardService.getBoardList();
+		if (boardList.size() > 0) {
+			boardSeq = boardList.get(0).getBoard_seq();
+		}
+
+		logger.info("boardSeq:[{}]", boardSeq);
 	}
 
-    /** 게시글 상세 조회 시 NULL이 아니면 테스트 통과 */
+	/** 게시글 목록 조회 시 NULL 아니면 테스트 통과 */
 	@Test
-	public void testGetBoardDetail() {
-		
+	public void testGetBoardList() {
+
 		try {
-			
-			List<Board> bookList = boardService.getBoardList();
-			int boardSeq = bookList.get(0).getBoard_seq();
-			
-			Board bookDetail = boardService.getBoardDetail(boardSeq);
-			assertNotNull(bookDetail);
-			
+
+			List<Board> boardList = boardService.getBoardList();
+			assertNotNull(boardList);
+
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
 	}
 
+	/** 게시글 상세 조회 시 NULL이 아니면 테스트 통과 */
+	@Test
+	public void testGetBoardDetail() {
+
+		try {
+
+			if (boardSeq != 0) {
+
+				Board boardDetail = boardService.getBoardDetail(boardSeq);
+				assertNotNull(boardDetail);
+			}
+
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+	}
+
+	/** 게시글 등록 후 1이 응답되면 테스트 통과 */
 	@Rollback(true)
 	@Test
 	public void testInsertBoard() {
-		
+
 		try {
 
 			Board board = new Board();
 			board.setBoard_writer("게시글 작성자 등록");
 			board.setBoard_subject("게시글 제목 등록");
-			board.setBoard_content("게시글 내용 등록");			
-			
-			int result = boardService.insertBoard(board);			
+			board.setBoard_content("게시글 내용 등록");
+
+			int result = boardService.insertBoard(board);
+
 			assertTrue(result == 1);
 
 		} catch (Exception e) {
 			e.printStackTrace();
-		}		
+		}
 	}
 
 	/** 게시글 수정 후 1이 응답되면 테스트 통과 */
 	@Rollback(true)
 	@Test
 	public void testUpdateBoard() {
-		
-		try {
-			
-			List<Board> bookList = boardService.getBoardList();
-			int boardSeq = bookList.get(0).getBoard_seq();
 
-			Board board = new Board();
-			board.setBoard_seq(boardSeq);
-			board.setBoard_subject("게시글 제목 수정");
-			board.setBoard_content("게시글 내용 수정");
-			
-			int result = boardService.updateBoard(board);
-			assertTrue(result == 1);
+		try {
+
+			if (boardSeq != 0) {
+
+				Board board = new Board();
+				board.setBoard_seq(boardSeq);
+				board.setBoard_subject("게시글 제목 수정");
+				board.setBoard_content("게시글 내용 수정");
+
+				int result = boardService.updateBoard(board);
+				assertTrue(result == 1);
+			}
 
 		} catch (Exception e) {
 			e.printStackTrace();
@@ -105,12 +123,12 @@ public class BoardServiceTest {
 	public void testDeleteBoard() {
 
 		try {
-			
-			List<Board> bookList = boardService.getBoardList();
-			int boardSeq = bookList.get(0).getBoard_seq();
-			
-			int result = boardService.deleteBoard(boardSeq);			
-			assertTrue(result == 1);
+
+			if (boardSeq != 0) {
+
+				int result = boardService.deleteBoard(boardSeq);
+				assertTrue(result == 1);
+			}
 
 		} catch (Exception e) {
 			e.printStackTrace();
